Внесены изменения по замечаниям

+++ comics_about_Python/post_vk.py
@@ -2,23 +2,26 @@ import os, requests, json_file, fetch_xckd, random, argparse
 from dotenv import load_dotenv
 from pprint import pprint
 
-def get_groups_admin(access_token):
-    url='https://api.vk.com/method/groups.get?user_id=166113106&filter=admin&access_token={}&v=5.95'.format(access_token)
+def get_vk(method, payload):
+    url = 'https://api.vk.com/method/{}?v=5.95'.format(method)
+    response = requests.get(url, params=payload)
 
-    response = requests.get(url)
-    if response.ok:
-        pprint(response.json())
+    return response.json().get('response')
 
-def get_server_address_to_upload_photos(group_id, access_token):
-    url = 'https://api.vk.com/method/photos.getWallUploadServer?group_id={}&access_token={}&v=5.95'.format(group_id, access_token)
 
-    response = requests.get(url)
-    if response.ok:
-        upload_url = response.json()['response']['upload_url']
-    return upload_url
+def get_server_address_to_upload_photos(access_token, group_id):
+    method = 'photos.getWallUploadServer'
+    payload = {
+        'group_id':group_id,
+        'access_token':access_token,
+    }
+
+    response = get_vk(method, payload)
+    return None if response is None else response['upload_url']
+
 
 def upload_photo(upload_url, directory, filename, caption):
-    filepath = '{}/{}'.format(directory, filename)
+    filepath = os.path.join(directory, filename)
 
     files = {
         'photo': open(filepath, 'rb'),
@@ -27,23 +30,36 @@ def upload_photo(upload_url, directory, filename, caption):
             }
     response = requests.post(upload_url, files=files)
 
-    return response.json()['hash'], response.json()['server'], response.json()['photo']
+    return (None, None, None) if response is None else (response.json()['hash'], response.json()['server'], response.json()['photo'])
 
 
-def save_photo(user_id, group_id, photo, hash, server, caption, access_token):
-    url='https://api.vk.com/method/photos.saveWallPhoto?user_id={}&group_id={}&photo={}&hash={}&server={}&caption={}&access_token={}&v=5.95'.format(user_id, group_id, photo, hash, server, caption, access_token)
-    response = requests.get(url)
-    return 'photo'+str(response.json()['response'][0]['owner_id'])+'_'+str(response.json()['response'][0]['id'])
+def save_photo(access_token, user_id, group_id, photo, hash, server, caption):
+    method = 'photos.saveWallPhoto'
+    payload = {
+        'user_id':user_id,
+        'group_id':group_id,
+        'photo':photo,
+        'hash':hash,
+        'server':server,
+        'caption':caption,
+        'access_token':access_token,
+    }
+
+    response = get_vk(method, payload)
+    return None if response is None else 'photo'+str(response[0]['owner_id'])+'_'+str(response[0]['id'])
 
 
-def post_photo(message, group_id, attachments, access_token):
-    from_group = '1'
-    owner_id = '-' + group_id
+def post_on_wall(access_token, message, group_id, attachments=''):
+    method = 'wall.post'
+    payload = {
+        'owner_id':'-' + group_id,
+        'from_group': '1',
+        'message':message,
+        'attachments':attachments,
+        'access_token':access_token,
+    }
 
-    url='https://api.vk.com/method/wall.post?owner_id={}&from_group={}&attachments={}&message={}&access_token={}&v=5.95'.format(owner_id, from_group, attachments, message, access_token)
-    response = requests.get(url)
-    if response.ok:
-        pprint(response.json())
+    get_vk(method, payload)
 
 
 def post_image(directory, json_filename, image_id):
@@ -55,18 +71,24 @@ def post_image(directory, json_filename, image_id):
     filename = file_contents[image_id]['filename']
     caption = file_contents[image_id]['description']
     title = file_contents[image_id]['title']
-    filepath = '{}/{}'.format(directory, filename)
+    filepath = os.path.join(directory, filename)
 
-    upload_url = get_server_address_to_upload_photos(group_id, access_token)
+    upload_url = get_server_address_to_upload_photos(access_token, group_id)
+    if upload_url is None:
+        return False
     hash, server, photo = upload_photo(upload_url, directory, filename, caption)
-    attachments = save_photo(user_id, group_id, photo, hash, server, caption, access_token)
-    post_photo(title, group_id, attachments, access_token)
+    if hash is None or server is None or photo is None:
+        return False
+    attachments = save_photo(access_token, user_id, group_id, photo, hash, server, caption)
+    if attachments is None:
+        return False
+    post_on_wall(access_token, title, group_id, attachments)
 
     os.remove(filepath)
 
     file_contents[image_id]["posted"] = True
     json_file.write_file(directory, json_filename, file_contents)
-
+    return True
 
 def get_random_image_id(directory, json_filename, max_image_id):
     file_contents = json_file.load_file(directory, json_filename)
@@ -90,21 +112,22 @@ def post_xckd_comics(image_info=None):
         image_id = str(max_image_id)
     elif image_info.upper() == 'RANDOM':
         image_id = get_random_image_id(directory, json_filename, max_image_id)
+    elif not image_info.isdigit():
+        return False
     elif int(image_info) > int(max_image_id):
         return False
     else:
         image_id = image_info
 
     file_contents = json_file.load_file(directory, json_filename)
-    if image_id in file_contents:
-        if not file_contents[image_id]['posted']:
-            post_image(directory, json_filename, image_id)
-        else:
-            return False
-    else:
+    if image_id not in file_contents:
         fetch_xckd.fetch_xckd_comics(directory, json_filename, image_id)
-        post_image(directory, json_filename, image_id)
-    return True
+        return post_image(directory, json_filename, image_id)
+
+    if not file_contents[image_id]['posted']:
+        return post_image(directory, json_filename, image_id)
+    else:
+        return False
 
 
 def main():
